backend/requirements.txt
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
Deprecated==1.2.14
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-PyMongo==2.3.0
flask-redis==0.4.0
Flask-Testing==0.8.1
gunicorn==21.2.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ndg-httpsclient==0.5.1
ordered-set==4.1.0
packaging==23.2
pyasn1==0.5.1
pycparser==2.21
Pygments==2.17.2
pymongo==3.11.4
pyOpenSSL==24.0.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
rich==13.7.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
Deprecated==1.2.14
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-PyMongo==2.3.0
flask-redis==0.4.0
Flask-Testing==0.8.1
gunicorn==21.2.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ndg-httpsclient==0.5.1
ordered-set==4.1.0
packaging==23.2
pyasn1==0.5.1
pycparser==2.21
Pygments==2.17.2
pymongo==3.11.4
pyOpenSSL==24.0.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
rich==13.7.0
typing_extensions==4.9.0
urllib3==2.1.0
uWSGI==2.0.24
Werkzeug==3.0.1
wrapt==1.16.0
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
Deprecated==1.2.14
dnspython==2.6.0
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-PyMongo==2.3.0
flask-redis==0.4.0
Flask-Testing==0.8.1
gunicorn==21.2.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ndg-httpsclient==0.5.1
ordered-set==4.1.0
packaging==23.2
pyasn1==0.5.1
pycparser==2.21
Pygments==2.17.2
pymongo==3.11.4
pyOpenSSL==24.0.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
rich==13.7.0
typing_extensions==4.9.0
urllib3==2.1.0
uWSGI==2.0.24
Werkzeug==3.0.1
wrapt==1.16.0
